Fix compiler warnings (#70)
This commit fixes all of the compile errors that we see in our versions
of gcc and clang. Also, we have refactored all of the for loops in the
source code to either use the new iterator-style syntax or to use
`decltype()` to determine the proper type for the index variable.