Make ZeroMQ optional (#74)
This commit adds a USE_ZMQ flag to cmake and re-works the build
to honor the flag.  As a result, ZeroMQ isn't configured globally
as part of the context anymore; rather, it's initialized as part
of the stream_publisher processor.  This means that even if you
build with USE_ZMQ, streamer pipelines won't publish themselves
via ZeroMQ unless you explicitly use the stream_published. The
ZeroMQ port is also now configured as part of that processor
instead of globally via a config file.