Add support for extracting intermediate layer activations (#55)
This very large commit adds support for extracting intermediate
activations from Caffe neural net models. In order to accomplish this in
the most extensible way, we have added a new type of Processor called
NeuralNetEvaluator that is solely responsible for evaluating a neural
network.

Follow-on tasks like assigning labels based on the output of a neural
net are accomplished by routing the output of a NeuralNetEvaluator into
a lightweight Processor that only performs the post processing. These
post processing Processors inherit from an additional new type of
Processor, the NeuralNetConsumer, which can automatically create and
manage a NeuralNetEvaluator. We have modified ImageClassifier to adhere
to this architecture.

A NeuralNetConsumer can exist in two modes, depending on whether the
NeuralNetEvaluator that it follows was explicitly created by the user
(explicit mode) or if it was automatically created by NeuralNetConsumer
(hidden mode). Users should explicitly create a NeuralNetEvaluator if
they would like to share its results between several NeuralNetConsumers.

In order to simplify the implementation of this change, we removed the
broken batching support from ImageClassifier. This required modifying
the multicam and benchmark apps, and it may break other user apps.