Refactor Processor::GetType() (#81)
Previously, Processor::GetType() was overriden by each Processor
subclass. This was excessive given that it always simply returned a
hardcoded ProcessorType. This commit adds a ProcessorType data member to
the Processor class, implements GetType() in Processor.cpp, and removes
all overriding declarations of GetType(). Processor subclasses pass
their ProcessorType to the Processor constructor, where it is stored in
the abovementioned data member for retrieval later.