Move ProcessorFactory functionality into Processor classes (#80)
This commit moves all of the functionality from the
ProcessorFactory::Create*() functions into their respective Processors'
classes. This means that all Processor creation logic is now contained
within the Processor classes themselves, which improves modularity and
reduces the number of #ifdef's in ProcessorFactory.

Since this commit requires adding additional #include statements to most
Processors, I use this opportunity to fix improper #include formatting.

This code is largely untested (as was the existing ProfessorFactory
code). This will be remedied shortly when the new JSON user api is
implemented.

This commit resolves Issue #45.