Remove default case from ProcessorFactory
In switch statements that are conditioned on an enum, having a default
case suppresses compiler warnings about whether all values of the enum
have associated case statements. In order to make sure that these
warnings are printed, this commit removes the default case from
ProcessorFactory::Create(). Instead, if the control flow falls through
the case statement, then a fatal error is triggered.