Fix bug that prevented GSTCamera from stopping
Instead of blocking forever when waiting for a new frame,
GstVideoCapture::GetPixels() now times out after 100 milliseconds. This
makes it possible for the GSTCamera processor to be stopped in the case
that is it reading a file from disk and has reached the end of the file.
Previously, in this case GstVideoCapture::GetPixels() would have been
blocked waiting for new pixels, preventing GSTCamera from stopping.