Add Frame field for capture time (in microseconds) (#118)
Camera::MetadataToFrame() now adds the current time in microseconds to
each frame with the key "capture_time_micros". We are using Boost's
date_time library instead of std::chrono for ease of serialization.

This change means that MetadataToFrame() is assumed to be called by all
of the camera subclasses whenever a frame is created. Therefore, for
simplicity, this commit also movies the creation of the "frame_id"
field into MetadataToFrame().