Refactor FileWriter processor and serializer app (#122)
The FileWriter has been renamed to FrameWriter, since it writes
whole Frames. This change is to differentiate it from the JpegWriter
and BinaryFileWriter, which operate on single Frame fields.

The serializer app is now the frame_writer app, in order to keep with the
naming convention for the various writing apps.

Previously, the FileWriter saved all frames in the same file. This made
parsing the frame data (both manually and programmatically) very
difficult. Therefore, this commit changes the FileWriter to save each
frame in an individual file.

Additionally, the FrameWriter now supports saving a subset of frame
fields.