Fix summary output for Caffe
The summary used to print the location of CAFFE_HOME, which is
a user-specified variable.  If you have Caffe installed in a
standard location (e.g., /usr/include), then you don't have to
set CAFFE_HOME at all--cmake will find it just fine.  In such
a case, the cmake summary wouldn't print anything for Caffe;
this patch fixes that as the Caffe_INCLUDE_DIRS is set whether
or not the user sets CAFFE_HOME.