Add blocking queue mode to processor class
Each processor can now be configured to block when pushing frames to its
output queues, as opposed to dropping a frame if an output queue is
full. When applied to all processors, this will entirely prevent frames
from being dropped due to queue overflows. This is particularly useful
when running offline experiments where frame drops are undesirable.