Send stop frame through pipeline
This commit adds functionality to determine when a pipeline is done
processing all of the frames in a file. This required modifying
GSTCamera and GstVideoCapture to detect the last frame in a stream.
Since this new feature only applies to reading files, the GSTCamera is
the only camera that currently supports detecting the end of a stream.

To propogate this information through the pipeline, the camera code adds
a "stop_frame" entry to the fields map of an empty frame and forwards
the frame down the pipeline. When a processor detects a frame that
contains the "stop_frame" field, it does not process it and just
immediately forwards it to all of its sinks.

When the user application detects a stop frame, it can be sure that it
will receive no more meaningful data frames from the corresponding
StreamReader. The pipeline does not automatically stop itself due to the
complexities of calling Processor::Stop() from the process thread.
Instead, the user application is responsible for detecting the stop
frame and stopping the pipeline itself. The "simple" app demonstrates
this.