Fix frame-popping bug
In gst_video_capture.cpp, raw frame data is retrieved from the head of a
queue, but items were being removed from the back of the queue. If the
queue ever contained more than one element, then subsequent frames would
contain the same raw data. This commit changes the logic to remove
frames from the front of the queue.