Fix race condition in Processor::ProcessorLoop()
A check would fail if, after forwarding a stop frame, the corresponding
process thread entered its next iteration before the application thread
called Processor::Stop(). This occured because
Processor::ProcessorLoop() was incorrectly ignoring null frames. This
commit changes Processor::ProcessorLoop() to immediately return if it
detects a null frame, since now the only way for a null frame to be
retrieved is if the processor is in the process of stopping.