Refactor train detector apps to drop fields
This commit adds functionality to train_detector_publisher.cpp,
train_detector_subscriber.cpp, frame_publisher.cpp, and frame_writer.cpp
to enable dropping of unwanted fields. The goal of this is to avoid the
serialization/deserialization and network bandwidth costs of extraneous
fields so that the train detector apps can run on machines with limited
CPU resources and network bandwidth.