Introduce OutputTracker for creating dir trees
The OutputTracker is useful for processors that save output data and
would like to either control the number of frames in each output
subdirectory or save frames in a hierarchy based on their capture
timestamp. This commit refactors FrameWriter and JpegWriter to take
advantage of OutputTracker and removes the JPEG functionality from
FrameWriter.