Apply backpressure to GStreamer pipeline
This commit modifies the GstVideoCapture class to automatically apply
backpressure to its GStreamer pipeline if its internal frame queue grows
past a user-defined threshold. This occurs if the GStreamer pipeline is
producing frames faster than the Streamer pipeline can consume them.
This primarily occurs when a GSTCamera is set to block when pushing
frames (e.g. when reading a file). If the GSTCamera is not set to block
when pushing frames, then it consumes frames as fast as it can and any
drops induced by a slow Streamer pipeline are handled by the inter-
Processor queues (StreamReaders).

The threshold is specified on a per-camera basis in "cameras.toml" by
the "max_buf_size" parameter, which must be greater than 1 and has a
default value of 10.

The backpressure feature only apply to GSTCameras (i.e. cameras
specified as "gst://", "rtsp://", or "file://").