Bug fix for commit f3fb6b7
This commit fixes a bug in the following commit:

    commit f3fb6b78fdc24179058c3e81ff6e475d602d3c59
    Author: Christopher Canel <c.canel@icloud.com>
    Date:   Tue Sep 12 18:37:18 2017 -0400

        Make frame printing handle high-dimensionality matrices

That commit ended up serializing the entire frame array into a
string before truncating it for printing.  This commit restores
the original behavior which copies just the first few elements
into a temporary array before serialization.

The bug manifested as frame drops in the 'simple' app, which
prints frames to the terminal.