Make JpegWriter push frames
This commit makes the JpegWriter forward frames after writing the JPEG
file. This allows other processors to use the information stored in the
frame about where the JPEG is stored, and provides a mechanism for
enforcing that a frame's JPEG is saved before other processing takes
place, which is a guarantee that cannot be made if the JpegWriter is in
parallel with the other processing.