Add license boilerplate to source files
This commit adds Apache 2.0 license boilerplate to all source files that
support comments. In instances where a file was heavily influenced by a
file from another project, our license was added and the other project
was referenced in the following comment. In the single instance where a
file already had another license
(scripts/groundtruth-gen/classification.cpp), the existing license was
modified to account for changes made by the Streamer authors.

This commit also adds an AUTHORS file, which is referred to by the
license boilerplate when it references "The Streamer Authors".